[feat] Implement the audio explorer based on the base explorer
