+++ aim/web/ui/src/modules/BaseExplorer/components/AudioBox/index.tsx
@@ -1,0 +1,3 @@
+import AudioBox from './AudioBox';
+
+export default AudioBox;
